Experiment: Make destination filenames upper case
When copying local files they are renamed so that the original filename
is represented using upper case letters.

NOTE: This is an experiment and should not be merged into main!!

Implements a work-around for a "bug" that confuses Bernese when mixing
RINEX2 and RINEX3 data from the same station. This happens in particular
when the RINEX2 file has the station name written in lower case letters
and the RINEX3 file has the station name given in upper case letters.
Somehow Bernese interpretes that as two different stations. By forcing
the copied filenames to be all upper case we should avoid that.